+++ R/build-news.R
@@ -132,12 +132,12 @@ globalVariables(".")
 
 data_news <- function(pkg = ".") {
   pkg <- as_pkgdown(pkg)
-  scoped_file_context(depth = 1L)
 
   html <- markdown(path(pkg$src_path, "NEWS.md"))
+  xml <- xml2::read_html(html)
+  downlit::downlit_html_node(xml)
 
-  sections <- xml2::read_html(html) %>%
-    xml2::xml_find_all("./body/div")
+  sections <- xml2::xml_find_all(xml, "./body/div")
 
   titles <- sections %>%
     xml2::xml_find_first(".//h1|h2") %>%
@@ -161,7 +161,6 @@ data_news <- function(pkg = ".") {
   }
 
   html <- sections %>%
-    purrr::walk(tweak_code) %>%
     purrr::walk2(versions, tweak_news_heading, timeline = timeline) %>%
     purrr::map_chr(as.character) %>%
     purrr::map_chr(repo_auto_link, pkg = pkg)