+++ tests/testthat/assets/reference-fail/R/f.R
@@ -1,0 +1,7 @@
+#' Title
+#'
+#' \url{}
+#' @export
+f <- function() {
+
+}